Give the two duplicated test fixtures one home each
test_mqtt and test_rest carried forty-six byte-for-byte identical lines: a fake
clock, setUp/tearDown, makeBridge() with five literal bridge values, and the
whole Rig that drives the real EverSolar driver against the simulated inverter.
The literals are what made this worth extracting. Both suites assert on the same
payload builders reading the same BridgeInfo, so a drifted uptime or firmware
version in one copy would have had the two suites describing different bridges,
with nothing to say which one was right.

Proved load-bearing rather than assumed: changing uptimeSeconds in the shared
fixture turns BOTH suites red (3 failures across test_mqtt and test_rest).

The fake clock moves with it but keeps the name `g_now`, pulled in by a
using-declaration. The two suites reference it about ninety times; renaming those
to gain a shared declaration would have been churn, not cleanup. For the same
reason the other six suites that declare their own two-line clock are left alone
-- each test binary is independent, so those copies cannot drift into each other,
and a shared header there would add a file to save five lines.

Modbus RTU's trailing CRC was appended inline in thirteen places across three
suites, in two spellings that differ only in a redundant mask on the high byte.
A checksum whose byte order is written out thirteen times is thirteen chances to
write it backwards once -- and a test that builds a frame backwards and asserts
against a decoder reading it backwards passes. Mutating the helper to emit the
high byte twice fails 21 cases across all three suites.

Left inline deliberately: the three sites that push a DELIBERATELY wrong CRC
(`^ 0xFF`), where the byte-level spelling is the point of the test, and
test_modbus/rtu.cpp, which is the suite that tests crc16 itself -- a helper built
on the thing under test would assert nothing.

The helper lives in heliograph::test, not a global `test`. A second top-level
namespace of that name is ambiguous the moment a suite says
`using namespace heliograph;`, which is how the first attempt failed to compile.
